+++ packages/utils/auth/index.ts
@@ -1,2 +1,2 @@
-export * from './clerk-types'
+export * from './auth-types'
 export * from './useAuthState'